+++ src/common/init.c
@@ -117,6 +117,7 @@ static void defaultConfig(FFinstance* instance)
     initModuleArg(&instance->config.users);
     initModuleArg(&instance->config.bluetooth);
     initModuleArg(&instance->config.sound);
+    initModuleArg(&instance->config.gamepad);
 
     ffStrbufInitA(&instance->config.libPCI, 0);
     ffStrbufInitA(&instance->config.libVulkan, 0);
@@ -164,7 +165,7 @@ static void defaultConfig(FFinstance* instance)
 
     instance->config.bluetoothShowDisconnected = false;
 
-    instance->config.soundShowAll = false;
+    instance->config.soundType = FF_SOUND_TYPE_MAIN;
 
     ffStrbufInitA(&instance->config.batteryDir, 0);
 
@@ -173,7 +174,9 @@ static void defaultConfig(FFinstance* instance)
     instance->config.localIpShowIpV4 = true;
     instance->config.localIpShowIpV6 = false;
     instance->config.localIpShowLoop = false;
+    instance->config.localIpV6First = false;
     ffStrbufInit(&instance->config.localIpNamePrefix);
+    instance->config.localIpCompactType = FF_LOCALIP_COMPACT_TYPE_NONE;
 
     instance->config.publicIpTimeout = 0;
     ffStrbufInit(&instance->config.publicIpUrl);
@@ -385,6 +388,8 @@ static void destroyConfig(FFinstance* instance)
     destroyModuleArg(&instance->config.openCL);
     destroyModuleArg(&instance->config.users);
     destroyModuleArg(&instance->config.bluetooth);
+    destroyModuleArg(&instance->config.sound);
+    destroyModuleArg(&instance->config.gamepad);
 
     ffStrbufDestroy(&instance->config.libPCI);
     ffStrbufDestroy(&instance->config.libVulkan);