+++ src/fastfetch.h
@@ -34,6 +34,20 @@ typedef enum FFLogoType
     FF_LOGO_TYPE_NONE,        //--logo none
 } FFLogoType;
 
+typedef enum FFSoundType
+{
+    FF_SOUND_TYPE_MAIN,
+    FF_SOUND_TYPE_ACTIVE,
+    FF_SOUND_TYPE_ALL,
+} FFSoundType;
+
+typedef enum FFLocalIpCompactType
+{
+    FF_LOCALIP_COMPACT_TYPE_NONE,
+    FF_LOCALIP_COMPACT_TYPE_MULTILINE,
+    FF_LOCALIP_COMPACT_TYPE_ONELINE,
+} FFLocalIpCompactType;
+
 typedef enum FFBinaryPrefixType
 {
     FF_BINARY_PREFIX_TYPE_IEC,   // 1024 Bytes = 1 KiB, 1024 KiB = 1 MiB, ... (standard)
@@ -141,6 +155,7 @@ typedef struct FFconfig
     FFModuleArgs users;
     FFModuleArgs bluetooth;
     FFModuleArgs sound;
+    FFModuleArgs gamepad;
 
     FFstrbuf libPCI;
     FFstrbuf libVulkan;
@@ -193,9 +208,11 @@ typedef struct FFconfig
     FFstrbuf separatorString;
 
     bool localIpShowLoop;
+    bool localIpV6First;
     bool localIpShowIpV4;
     bool localIpShowIpV6;
     FFstrbuf localIpNamePrefix;
+    FFLocalIpCompactType localIpCompactType;
 
     FFstrbuf publicIpUrl;
     uint32_t publicIpTimeout;
@@ -203,7 +220,7 @@ typedef struct FFconfig
     FFstrbuf weatherOutputFormat;
     uint32_t weatherTimeout;
 
-    bool soundShowAll;
+    FFSoundType soundType;
 
     FFstrbuf osFile;
 
@@ -319,5 +336,6 @@ void ffPrintUsers(FFinstance* instance);
 void ffPrintCommand(FFinstance* instance);
 void ffPrintBluetooth(FFinstance* instance);
 void ffPrintSound(FFinstance* instance);
+void ffPrintGamepad(FFinstance* instance);
 
 #endif